+++ src/Array/__tests__/concat.spec.lua
@@ -6,8 +6,8 @@ return function()
 	local concat = require(Array.concat)
 
 	local Packages = LuauPolyfill.Parent
-	local JestRoblox = require(Packages.Dev.JestRoblox)
-	local jestExpect = JestRoblox.Globals.expect
+	local JestGlobals = require(Packages.Dev.JestGlobals)
+	local jestExpect = JestGlobals.expect
 
 	it("concatenate arrays with single values", function()
 		jestExpect(concat({ 1 })).toEqual({ 1 })