+++ src/Array/__tests__/every.spec.lua
@@ -7,8 +7,8 @@ return function()
 	local every = require(Array.every)
 
 	local Packages = LuauPolyfill.Parent
-	local JestRoblox = require(Packages.Dev.JestRoblox)
-	local jestExpect = JestRoblox.Globals.expect
+	local JestGlobals = require(Packages.Dev.JestGlobals)
+	local jestExpect = JestGlobals.expect
 
 	it("Invalid argument", function()
 		jestExpect(function()