+++ src/Array/__tests__/filter.spec.lua
@@ -8,8 +8,8 @@ return function()
 	local isFinite = require(LuauPolyfill.Number.isFinite)
 
 	local Packages = LuauPolyfill.Parent
-	local JestRoblox = require(Packages.Dev.JestRoblox)
-	local jestExpect = JestRoblox.Globals.expect
+	local JestGlobals = require(Packages.Dev.JestGlobals)
+	local jestExpect = JestGlobals.expect
 
 	it("Filtering out all small values", function()
 		local isBigEnough = function(value)