+++ src/Array/__tests__/find.spec.lua
@@ -4,8 +4,8 @@ return function()
 	local find = require(Array.find)
 
 	local Packages = LuauPolyfill.Parent
-	local JestRoblox = require(Packages.Dev.JestRoblox)
-	local jestExpect = JestRoblox.Globals.expect
+	local JestGlobals = require(Packages.Dev.JestGlobals)
+	local jestExpect = JestGlobals.expect
 
 	local function returnTrue()
 		return true