+++ src/Array/__tests__/map.spec.lua
@@ -6,8 +6,8 @@ return function()
 	local map = require(Array.map)
 
 	local Packages = LuauPolyfill.Parent
-	local JestRoblox = require(Packages.Dev.JestRoblox)
-	local jestExpect = JestRoblox.Globals.expect
+	local JestGlobals = require(Packages.Dev.JestGlobals)
+	local jestExpect = JestGlobals.expect
 
 	it("Invalid argument", function()
 		-- roblox-cli analyze fails because map is called with an