+++ src/Object/__tests__/values.spec.lua
@@ -4,8 +4,8 @@ return function()
 
 	local LuauPolyfill = Object.Parent
 	local Packages = LuauPolyfill.Parent
-	local JestRoblox = require(Packages.Dev.JestRoblox)
-	local jestExpect = JestRoblox.Globals.expect
+	local JestGlobals = require(Packages.Dev.JestGlobals)
+	local jestExpect = JestGlobals.expect
 
 	it("returns the values of a table", function()
 		local result = values({