+++ src/Timers/__tests__/Timers.spec.lua
@@ -5,8 +5,8 @@ return function()
 	local createSpy = require(Timers.Parent.createSpy)
 
 	local Packages = LuauPolyfill.Parent
-	local JestRoblox = require(Packages.Dev.JestRoblox)
-	local jestExpect = JestRoblox.Globals.expect
+	local JestGlobals = require(Packages.Dev.JestGlobals)
+	local jestExpect = JestGlobals.expect
 
 	local Timeout
 	local mockTime, timeouts